Refactor C-style casts to modern C++ casts (static_cast)
Replaced legacy C-style casts (e.g., `(int)val`) with `static_cast<type>(val)` across multiple modules to improve type safety and comply with modern C++ standards.

Changes include:
- `source/editor/operations/clean_operations.h`: Casts in UNREACHABLE_SEARCH_RADIUS calculations.
- `source/ui/replace_tool/rule_list_control.cpp`: Casts in list indexing and size comparisons.
- `source/ui/replace_tool/visual_similarity_service.cpp`: Casts in RGB binning and coordinate calculations.
- `source/ui/replace_tool/rule_builder_panel.cpp`: Casts in rule management and layout logic.
- `source/ui/replace_tool/rule_card_renderer.cpp`: Casts in renderer layout calculations.
- `source/game/item.cpp`: Cast in weight calculation.
- `source/rendering/drawers/overlays/preview_drawer.cpp`: Cast for house ID comparison and drawing coordinates.
- `source/rendering/drawers/minimap_drawer.cpp`: Casts for GL coordinates and dimensions.
- `source/rendering/drawers/entities/creature_name_drawer.cpp`: Casts for screen position calculations.
- `source/rendering/drawers/minimap_renderer.cpp`: Casts for instance data and scaling factors.

+++ source/ui/replace_tool/rule_builder_panel.cpp
@@ -50,16 +50,16 @@ bool RuleBuilderPanel::ItemDropTarget::OnDropText(wxCoord x, wxCoord y, const wx
 	RuleBuilderPanel::HitResult hit = m_panel->HitTest(x, y);
 
 	if (hit.type == RuleBuilderPanel::HitResult::Source) {
-		if (hit.ruleIndex >= 0 && hit.ruleIndex < (int)m_panel->m_rules.size()) {
+		if (hit.ruleIndex >= 0 && hit.ruleIndex < static_cast<int>(m_panel->m_rules.size())) {
 			m_panel->m_rules[hit.ruleIndex].fromId = itemId;
 			m_panel->m_listener->OnRuleChanged();
 			m_panel->Refresh();
 			return true;
 		}
 	} else if (hit.type == RuleBuilderPanel::HitResult::Target || hit.type == RuleBuilderPanel::HitResult::DeleteTarget) {
-		if (hit.ruleIndex >= 0 && hit.ruleIndex < (int)m_panel->m_rules.size()) {
+		if (hit.ruleIndex >= 0 && hit.ruleIndex < static_cast<int>(m_panel->m_rules.size())) {
 			// Replace existing target
-			if (hit.targetIndex >= 0 && hit.targetIndex < m_panel->m_rules[hit.ruleIndex].targets.size()) {
+			if (hit.targetIndex >= 0 && hit.targetIndex < static_cast<int>(m_panel->m_rules[hit.ruleIndex].targets.size())) {
 				m_panel->m_rules[hit.ruleIndex].targets[hit.targetIndex].id = itemId;
 				m_panel->m_listener->OnRuleChanged();
 				m_panel->Refresh();
@@ -67,7 +67,7 @@ bool RuleBuilderPanel::ItemDropTarget::OnDropText(wxCoord x, wxCoord y, const wx
 			}
 		}
 	} else if (hit.type == RuleBuilderPanel::HitResult::AddTarget) { // Drag to [+] slot
-		if (hit.ruleIndex >= 0 && hit.ruleIndex < (int)m_panel->m_rules.size()) {
+		if (hit.ruleIndex >= 0 && hit.ruleIndex < static_cast<int>(m_panel->m_rules.size())) {
 			// TRASH LOGIC: Reject if has trash
 			bool hasTrash = std::ranges::any_of(m_panel->m_rules[hit.ruleIndex].targets, [](const auto& t) {
 				return t.id == TRASH_ITEM_ID;
@@ -156,7 +156,7 @@ std::vector<ReplacementRule> RuleBuilderPanel::GetRules() const {
 // ----------------------------------------------------------------------------
 
 int RuleBuilderPanel::GetRuleHeight(int index, int width) const {
-	if (index < 0 || index >= (int)m_rules.size()) {
+	if (index < 0 || index >= static_cast<int>(m_rules.size())) {
 		return FromDIP(ITEM_H);
 	}
 
@@ -166,7 +166,7 @@ int RuleBuilderPanel::GetRuleHeight(int index, int width) const {
 		availableWidth = CARD_W;
 	}
 
-	int columns = std::max(1, (int)(availableWidth / (CARD_W + ITEM_SPACING)));
+	int columns = std::max(1, static_cast<int>(availableWidth / (CARD_W + ITEM_SPACING)));
 	int targetCount = m_rules[index].targets.size();
 
 	bool hasTrash = std::ranges::any_of(m_rules[index].targets, [](const auto& t) {
@@ -182,8 +182,8 @@ int RuleBuilderPanel::GetRuleHeight(int index, int width) const {
 }
 
 int RuleBuilderPanel::GetRuleY(int index, int width) const {
-	if (index < 0 || index >= (int)m_ruleYCache.size()) {
-		if (index == (int)m_ruleYCache.size()) {
+	if (index < 0 || index >= static_cast<int>(m_ruleYCache.size())) {
+		if (index == static_cast<int>(m_ruleYCache.size())) {
 			return m_totalHeight;
 		}
 		return 0;
@@ -202,7 +202,7 @@ void RuleBuilderPanel::LayoutRules() {
 
 	for (size_t i = 0; i < m_rules.size(); ++i) {
 		m_ruleYCache.push_back(currentY);
-		currentY += GetRuleHeight((int)i, width);
+		currentY += GetRuleHeight(static_cast<int>(i), width);
 	}
 	m_totalHeight = currentY;
 
@@ -211,7 +211,7 @@ void RuleBuilderPanel::LayoutRules() {
 }
 
 void RuleBuilderPanel::DistributeProbabilities(int ruleIndex) {
-	if (ruleIndex < 0 || ruleIndex >= m_rules.size()) {
+	if (ruleIndex < 0 || ruleIndex >= static_cast<int>(m_rules.size())) {
 		return;
 	}
 	auto& targets = m_rules[ruleIndex].targets;
@@ -229,8 +229,8 @@ void RuleBuilderPanel::DistributeProbabilities(int ruleIndex) {
 
 	for (auto& target : targets) {
 		accumulated += step;
-		int currentTotal = (int)std::round(accumulated);
-		int prevTotal = (int)std::round(accumulated - step);
+		int currentTotal = static_cast<int>(std::round(accumulated));
+		int prevTotal = static_cast<int>(std::round(accumulated - step));
 		target.probability = currentTotal - prevTotal;
 	}
 }
@@ -332,8 +332,8 @@ RuleBuilderPanel::HitResult RuleBuilderPanel::HitTest(int x, int y) const {
 
 	// Check Rules
 	for (size_t i = 0; i < m_rules.size(); ++i) {
-		int ruleH = GetRuleHeight(i, width);
-		int ruleY = GetRuleY(i, width);
+		int ruleH = GetRuleHeight(static_cast<int>(i), width);
+		int ruleY = GetRuleY(static_cast<int>(i), width);
 		wxRect card(CARD_MARGIN_X, ruleY, width - CARD_MARGIN_X * 2, ruleH);
 
 		if (card.Contains(x, absY)) {
@@ -342,7 +342,7 @@ RuleBuilderPanel::HitResult RuleBuilderPanel::HitTest(int x, int y) const {
 
 			// Delete Rule Button (Top Right)
 			if (localX > card.width - 24 && localY < 24) {
-				return { HitResult::DeleteRule, (int)i, -1 };
+				return { HitResult::DeleteRule, static_cast<int>(i), -1 };
 			}
 
 			// Source Item (Left, always in first row logically)
@@ -350,13 +350,13 @@ RuleBuilderPanel::HitResult RuleBuilderPanel::HitTest(int x, int y) const {
 			float sourceY = CARD_PADDING; // Vertical top in card
 
 			if (localX >= startX && localX <= startX + CARD_W && localY >= sourceY && localY <= sourceY + ITEM_HEIGHT) {
-				return { HitResult::Source, (int)i, -1 };
+				return { HitResult::Source, static_cast<int>(i), -1 };
 			}
 
 			// Targets (Wrapping)
 			float targetStartX = startX + CARD_W + 10 + ARROW_WIDTH;
 			float availableWidth = card.width - targetStartX - CARD_PADDING;
-			int columns = std::max(1, (int)(availableWidth / (CARD_W + ITEM_SPACING)));
+			int columns = std::max(1, static_cast<int>(availableWidth / (CARD_W + ITEM_SPACING)));
 
 			for (size_t t = 0; t < m_rules[i].targets.size(); ++t) {
 				int row = t / columns;
@@ -365,7 +365,7 @@ RuleBuilderPanel::HitResult RuleBuilderPanel::HitTest(int x, int y) const {
 				float ty = RuleCardRenderer::CARD_PADDING + row * (ITEM_HEIGHT + RuleCardRenderer::ITEM_SPACING);
 
 				if (localX >= tx && localX <= tx + RuleCardRenderer::CARD_W && localY >= ty && localY <= ty + ITEM_HEIGHT) {
-					return { HitResult::DeleteTarget, (int)i, (int)t };
+					return { HitResult::DeleteTarget, static_cast<int>(i), static_cast<int>(t) };
 				}
 			}
 
@@ -382,11 +382,11 @@ RuleBuilderPanel::HitResult RuleBuilderPanel::HitTest(int x, int y) const {
 				float ty = RuleCardRenderer::CARD_PADDING + row * (ITEM_HEIGHT + RuleCardRenderer::ITEM_SPACING);
 
 				if (localX >= tx && localX <= tx + RuleCardRenderer::CARD_W && localY >= ty && localY <= ty + ITEM_HEIGHT) {
-					return { HitResult::AddTarget, (int)i, -1 };
+					return { HitResult::AddTarget, static_cast<int>(i), -1 };
 				}
 			}
 
-			return { HitResult::None, (int)i, -1 };
+			return { HitResult::None, static_cast<int>(i), -1 };
 		}
 	}
 
@@ -429,12 +429,12 @@ void RuleBuilderPanel::OnNanoVGPaint(NVGcontext* vg, int width, int height) {
 
 	// 2. Draw Content (Implicitly scrolled by base class)
 	for (size_t i = 0; i < m_rules.size(); ++i) {
-		int ruleY = GetRuleY(i, width);
-		bool hoverDel = (m_dragHover.type == HitResult::DeleteRule && m_dragHover.ruleIndex == (int)i);
-		int dragType = (m_dragHover.ruleIndex == (int)i) ? (int)m_dragHover.type : 0;
-		int dragIdx = (m_dragHover.ruleIndex == (int)i) ? m_dragHover.targetIndex : -1;
+		int ruleY = GetRuleY(static_cast<int>(i), width);
+		bool hoverDel = (m_dragHover.type == HitResult::DeleteRule && m_dragHover.ruleIndex == static_cast<int>(i));
+		int dragType = (m_dragHover.ruleIndex == static_cast<int>(i)) ? static_cast<int>(m_dragHover.type) : 0;
+		int dragIdx = (m_dragHover.ruleIndex == static_cast<int>(i)) ? m_dragHover.targetIndex : -1;
 
-		RuleCardRenderer::DrawRuleCard(this, vg, (int)i, ruleY, width, hoverDel, dragIdx, dragType, m_isExternalDrag);
+		RuleCardRenderer::DrawRuleCard(this, vg, static_cast<int>(i), ruleY, width, hoverDel, dragIdx, dragType, m_isExternalDrag);
 	}
 
 	int newRuleY = GetRuleY(m_rules.size(), width) + RuleCardRenderer::CARD_MARGIN_Y;
